requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
cairocffi==1.2.0
CairoSVG==2.5.0
cffi==1.14.4
cssselect2==0.4.1
defusedxml==0.6.0
Django==3.1.3
django-crispy-forms==1.9.2
django-filter==2.4.0
django-xhtml2pdf==0.0.4
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
Pillow==8.0.1
psycopg2-binary==2.8.6
pycparser==2.20
PyPDF2==1.26.0
Pyphen==0.10.0
python-bidi==0.4.2
pytz==2020.4
reportlab==3.5.55
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.0
WeasyPrint==52.1
webencodings==0.5.1
xhtml2pdf==0.2.5
